+++ agrest/src/main/java/io/agrest/meta/parser/EndpointMetadata.java
@@ -2,7 +2,6 @@ package io.agrest.meta.parser;
 
 import io.agrest.annotation.AgResource;
 import io.agrest.annotation.LinkType;
-import io.agrest.meta.annotation.Resource;
 
 import java.lang.reflect.Method;
 
@@ -12,24 +11,6 @@ import java.lang.reflect.Method;
 class EndpointMetadata {
 
     public static EndpointMetadata fromAnnotation(Method resourceMethod) {
-
-        EndpointMetadata mdLegacy = fromLegacyAnnotation(resourceMethod);
-        EndpointMetadata mdNew = fromNewAnnotation(resourceMethod);
-
-        if (mdLegacy != null && mdNew != null) {
-            throw new IllegalStateException("Method '" + resourceMethod.getName() + "' is annotated with both @Resource and @AgResource");
-        }
-
-        return mdNew != null ? mdNew : mdLegacy;
-    }
-
-    @Deprecated
-    private static EndpointMetadata fromLegacyAnnotation(Method resourceMethod) {
-        Resource annotation = resourceMethod.getAnnotation(Resource.class);
-        return annotation == null ? null : new EndpointMetadata(annotation.type().toType(), annotation.entityClass());
-    }
-
-    private static EndpointMetadata fromNewAnnotation(Method resourceMethod) {
         AgResource annotation = resourceMethod.getAnnotation(AgResource.class);
         return annotation == null ? null : new EndpointMetadata(annotation.type(), annotation.entityClass());
     }